requirements.txt
crewai>=0.11.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.5.0
typer>=0.9.0 